# Conflicts:
#	CHANGELOG.md
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/media/MediaService.kt
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/MediaCacheMapping.kt
#	trixnity-client/trixnity-client-repository-exposed/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/exposed/ExposedMediaCacheMappingRepository.kt
#	trixnity-client/trixnity-client-repository-realm/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/realm/RealmMediaCacheMappingRepository.kt
#	trixnity-client/trixnity-client-repository-room/src/commonMain/kotlin/net/folivo/trixnity/client/store/repository/room/RoomMediaCacheMappingRepository.kt

+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/MediaCacheMapping.kt
@@ -7,6 +7,6 @@ import kotlinx.serialization.Serializable
 data class MediaCacheMapping(
     val cacheUri: String,
     val mxcUri: String? = null,
-    val size: Int = 0,
+    val size: Long = 0,
     val contentType: String? = ContentType.Application.OctetStream.toString(),
 )
