+++ trixnity-core/src/commonMain/kotlin/net/folivo/trixnity/core/model/events/m/room/FileBasedInfo.kt
@@ -2,5 +2,5 @@ package net.folivo.trixnity.core.model.events.m.room
 
 interface FileBasedInfo {
     val mimeType: String?
-    val size: Int?
+    val size: Long?
 }
